chess-mcp-server/pyproject.toml
[project]
name = "chess-mcp-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.11.3",
    "python-a2a>=0.5.9",
    "python-chess>=1.999",
    "requests>=2.32.5",
    "stockfish>=3.28.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
]
